Create telegram_reminder
This scripts sets up automated reminders on a Telegram chat using Python. It leverages the telegram library and schedule the reminders using the schedule library. 

To use the script, replace "YOUR_BOT_TOKEN" with the token of your Telegram bot, and "YOUR_CHAT_ID" with the ID of the chat you want to send the reminder to. Also, replace "REMINDER MESSAGE" with the text of the reminder message you want to send.

Note that you need to create a Telegram bot and obtain its token in order to use the telegram library. You can create a new bot and obtain its token by following the instructions on the official Telegram documentation. You also need to add the bot to the chat you want to send the reminder to and obtain the chat ID by calling the bot.get_updates() method after the bot has been added to the chat.