Update whatsapp_message
This python script automates a message on WhatsApp every day at 8 am. It uses selenium library to interact with the WhatsApp web interface and the schedule library  to send it every day at 8 am. 

To use the code, replace "CHAT NAME" with the name of the chat you want to send the message to, and "MESSAGE TEXT" with the text of the message you want to send. Also, make sure to download the Chrome driver and set the path to the driver in the webdriver.Chrome() function.