Create lambda_whatsapp_notification
This python code sends automated messages on WhatsApp using the AWS Lambda SDK and S3. It creates a Lambda function that triggers when a new file is added to an S3 bucket.

To use this code, replace "YOUR_ACCOUNT_SID" and "YOUR_AUTH_TOKEN" with your Twilio account SID and auth token, respectively. Also, replace "whatsapp:+14155238886" with your Twilio WhatsApp number and "whatsapp:+WHATSAPP_NUMBER" with the phone number of the WhatsApp chat you want to send the message to.

Note that you need to set up a Twilio account and obtain a WhatsApp number in order to use the twilio library. You also need to configure the S3 bucket to trigger the Lambda function when a new file is added by using an S3 event notification.