Create twitter_media_download.py
This code can be used to retrieve all media files from a Twitter account using Python. To start, we need to import the Tweepy library which provides an easy-to-use interface to the Twitter API. 

Upon authentication with Twitter API using our consumer key, consumer secret, access token, and access token secret, we can create an API object and use it to retrieve the user's timeline using the user_timeline method. We specify the number of tweets to retrieve, whether to include retweets and replies, and the Twitter account name.

Next, we create a folder called "media" to store the media files and loop through each tweet in the timeline. If the tweet contains any media files, we download each file using the wget library and store it in the "media" folder.

Note that this code only retrieves the first 200 tweets from the user's timeline. If you want to retrieve more tweets, you can use the max_id parameter to get older tweets.