Create automating_tweets.py
To access Twitter's API, you need to create a developer account and get your API credentials. Once you have your credentials, you can use the tweepy library to interact with the Twitter API.

In the __init__ method, you pass your API credentials and initialize the tweepy.API object.

The post_tweet method takes a tweet as an argument and uses the api.update_status method to post the tweet to your Twitter account.

The get_tweets method takes a username and a count as arguments and returns a list of the user's tweets. It uses the tweepy.Cursor object to iterate over the user's timeline and collect the tweets.

To use this class, you need to create an instance of the TwitterAPI class and call its methods with your desired arguments.

Note that you should replace the consumer_key, consumer_secret, access_token, and access_token_secret arguments with your actual API credentials.